+++ changelog.md
@@ -14,6 +14,7 @@
 * Add new helper `action_link_to`
 * Add new helper `iconic_icon`
 * Add new helper `row_link_to`
+* Add new tabbable navigation helper `tabbable`
 
 ### Bugs Fixed
 * "class" option not being passed through button helpers